created discussion page and show all discussion record with channel name and establish a foreign key in discussion table

+++ database/migrations/2023_06_15_101056_create_discussions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('title');
             $table->text('content');
             $table->string('slug')->unique();
-            $table->integer('channel_id');
+            $table->unsignedBigInteger('channel_id');
             $table->timestamps();
         });
     }
